Fix TestCases
