+++ app-modules/integration-aws-ses-event-handling/tests/Landlord/Listeners/HandleSesEventTest.php
@@ -51,9 +51,7 @@ it('correctly handles the incoming SES event', function (string $event, EmailMes
     /** @var Tenant $tenant */
     $tenant = Tenant::query()->first();
 
-    $emailMessage = $tenant->execute(function () {
-        return EmailMessage::factory()->create();
-    });
+    $emailMessage = EmailMessage::factory()->create();
 
     // And we receive some sort of SES event when attempting to deliver
     $snsData = loadFixtureFromModule('integration-aws-ses-event-handling', 'sns-notification');
@@ -63,9 +61,7 @@ it('correctly handles the incoming SES event', function (string $event, EmailMes
     data_set($messageContent, 'mail.tags.tenant_id.0', $tenant->getKey());
     $snsData['Message'] = json_encode($messageContent);
 
-    $tenant->execute(function () use ($emailMessage) {
-        expect($emailMessage->events()->count())->toBe(0);
-    });
+    expect($emailMessage->events()->count())->toBe(0);
 
     $response = withHeaders(
         [
@@ -86,16 +82,14 @@ it('correctly handles the incoming SES event', function (string $event, EmailMes
 
     $response->assertOk();
 
-    $tenant->execute(function () use ($emailMessage, $eventType) {
-        // The email message should have the apppriate email message event created based on the event
-        $emailMessage->refresh();
+    // The email message should have the apppriate email message event created based on the event
+    $emailMessage->refresh();
 
-        expect($emailMessage->events()->count())->toBe(1);
+    expect($emailMessage->events()->count())->toBe(1);
 
-        $event = $emailMessage->events()->first();
+    $event = $emailMessage->events()->first();
 
-        expect($event->type)->toBe($eventType);
-    });
+    expect($event->type)->toBe($eventType);
 })->with([
     'HandleSesBounceEvent' => [
         'Bounce',
